requirements.txt
adal==1.2.7
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aiosignal==1.2.0
aiounittest==1.3.0
anyio==3.6.1
applicationinsights==0.11.9
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.28
Babel==2.7.0
backcall==0.2.0
bandit==1.7.4
beautifulsoup4==4.11.1
black==22.3.0
bleach==5.0.0
blessed==1.19.1
botbuilder-ai==4.14.2
botbuilder-applicationinsights==4.14.2
botbuilder-core==4.14.2
botbuilder-dialogs==4.14.2
botbuilder-integration-applicationinsights-aiohttp==4.14.2
botbuilder-schema==4.14.2
botframework-connector==4.14.2
botframework-streaming==4.14.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
click==8.1.3
colorful==0.5.4
confuse==1.7.0
coverage==6.4.1
cryptography==3.4.8
cycler==0.11.0
datatypes-date-time==1.0.0a2
datedelta==1.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==2.2.6
djangorestframework==3.10.3
emoji==1.7.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
flake8==4.0.1
Flask==1.1.1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-auth==2.8.0
googleapis-common-protos==1.56.2
gpustat==1.0.0b1
grapheme==0.6.0
graphviz==0.20
grpcio==1.43.0
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.1
importlib-metadata==4.11.4
iniconfig==1.1.1
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonpickle==1.4.2
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.17.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
missingno==0.5.1
mistune==0.8.4
modin==0.15.1
msal==1.6.0
msgpack==1.0.4
msrest==0.6.19
msrestazure==0.6.4
multidict==6.0.2
multipledispatch==0.6.0
mypy==0.961
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.22.4
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opencensus==0.9.0
opencensus-context==0.1.2
packaging==21.3
pandas==1.4.2
pandas-profiling==2.9.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pbr==5.9.0
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
plotly==5.8.2
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-spy==0.3.12
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.1.0
ray==1.12.1
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
redis==3.5.3
regex==2019.8.19
requests==2.23.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
smart-open==6.0.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
statsmodels==0.13.2
stevedore==3.5.0
tangled-up-in-unicode==0.2.0
tenacity==8.0.1
terminado==0.15.0
tinycss2==1.1.1
tokenize-rt==4.2.1
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
types-chardet==4.0.4
types-requests==2.27.30
types-urllib3==1.26.15
typing_extensions==4.2.0
urllib3==1.25.11
virtualenv==20.14.1
visions==0.5.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
